优化代码结构并且添加对日志的支持
