--- README.md
@@ -1,30 +1,0 @@
-# AutoXGP
-自动为您的微软账户订阅XGP、设置Minecraft档案、 取消订阅XGP
-
-您将获得一个可以通过Minecraft正版验证的账号
-
-## 使用
-* 在[config.ini](https://github.com/AneryCoft/AutoXGP/blob/main/config.ini)中完成对程序的设置
-
-如果订阅次数太多可能会被风控，这时候可能需要设置Proxy
-
-在设置多线程时请注意支付宝账户余额，每次订阅XGP价格为29港元
-
-如果需要自定义Minecraft皮肤，请设置Skin
-
-#### 注意：必须要对Alipay-payPassword进行配置
-
-* 在[account.txt](https://github.com/AneryCoft/AutoXGP/blob/main/account.txt)中写入账户
-
-
-如果您没有微软账户，请先[注册](https://signup.live.com/signup)
-
-#### 注意：账户的格式为 邮箱----密码 两个账户之间要换行
-
-* 运行[AutoXGP.py](https://github.com/AneryCoft/AutoXGP/blob/main/AutoXGP.py)
-完成后的XGP将会写入[XGP.txt](https://github.com/AneryCoft/AutoXGP/blob/main/XGP.txt)
-
-## 反馈
-如果您在运行过程的遇到报错，请将控制台截图到[issue](https://github.com/AneryCoft/AutoXGP/issue)
-
-QQ群：[225705430](http://qm.qq.com/cgi-bin/qm/qr?_wv=1027&k=HvEcIqMdYh5VScESIiN_vVjv7wO1n6zD&authKey=VikOm9rA7rMpGDNaNfH6%2BVP2SJRwTgLvmOYViAa2l6z%2FlLUE%2F10L%2FO8iXO5XFvMb&noverify=0&group_code=225705430)
