--- accounts.txt
@@ -1,1 +1,0 @@
-# Microsoft account format:Email----password
