--- config.ini
@@ -1,21 +1,0 @@
-[Proxy]
-enableProxy=false
-host=127.0.0.1
-port=0
-
-[Alipay]
-payPassword=
-saveCookie=true
-
-[Thread]
-thread=1
-
-[Prefix]
-xbox_prefix=AneryCoft
-minecraft_prefix=AneryCoft_
-
-[Skin]
-customSkin=false
-skin=skin.png
-model=0
-;0:经典 1:纤细
